extract an intrinsics.h
